eShopAnalysis.ETL/crawlProductCatalog/requirements.txt
ipython
pymongo
scrapy
selenium
jupyterlab
webdriver-manager
scrapy-selenium
python-slugify